Changed format of models and linted more
