+++ models/bert_model.py
@@ -1,0 +1,5 @@
+from spotifinders.models.nlp_model import NLPModel
+
+
+class BERTModel(NLPModel):
+
