+++ models/nlp_model.py
@@ -1,0 +1,7 @@
+from abc import ABC, abstractmethod
+
+
+class NLPModel(ABC):
+    @abstractmethod
+    def get_vector(self):
+        pass
